CMake Support

+++ include/Globals.h
@@ -5,7 +5,6 @@ typedef unsigned int u32;
 typedef int i32;
 
 #include <iostream>
-#include <thread>
 
 #include "Utility.h"
 
@@ -14,8 +13,8 @@ namespace modio
   struct CurrentDownloadInfo
   {
     std::string url;
-    long download_total;
-    long download_progress;
+    double download_total;
+	double download_progress;
   };
 
   extern std::string API_KEY;
